+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+2016-03-15 Benoit Jacquemont <benoit.jacquemont@gmail.com>
+* Added frame and symbol information
+* Added analyzers to find path that kept references
+
 2016-01-26 Benoit Jacquemont <benoit.jacquemont@gmail.com>
 * Added the object handle to uniquely identify objects
 
@@ -7,7 +11,5 @@
 2014-02-08 Benoit Jacquemont <benoit.jacquemont@gmail.com>
 * Migrating from Google Code to GitHub
 
-
 2011-01-16 Benoit Jacquemont <benoit.jacquemont@gmail.com>
-
 * Extension creation.